Add submenu into mobile nav
